requirements.txt
python-telegram-bot~=22.0
matplotlib~=3.10.1
dotenv~=0.9.9
python-dotenv~=1.1.0
pandas~=2.2.3
msoffcrypto-tool~=5.4.2
openpyxl~=3.1.5
SQLAlchemy~=2.0.40
alembic~=1.15.2
pillow~=11.1.0